Move AST factory helpers
