+++ tests/unit/services/mulang-test.js
@@ -1,6 +1,7 @@
 import QUnit, { module, test } from 'qunit'
 import { setupTest } from 'ember-qunit'
 import { blocklyWorkspaceMock } from '../../helpers/mocks'
+import { procedure, entryPoint, sequence, reference, application, repeat, muIf, ifElse, muWhile, number, string, none } from '../../helpers/astFactories'
 
 let pilasMulang = null
 
@@ -705,138 +706,3 @@ function mulangParseWorkspaceTest(description, definitions, expectedAst) {
   })
 }
 
-
-// TODO: Move to helpers
-// Builders
-function procedure(name, params, ...seq) {
-  return {
-    tag: "Procedure",
-    contents: [
-      name,
-      [
-        equation(params, ...seq)
-      ]
-    ]
-  }
-}
-
-function equation(params, ...seq) {
-  return [
-    params.map(name => variable(name)),
-    body(...seq)
-  ]
-}
-
-function variable(name) {
-  return {
-    tag: "VariablePattern",
-    contents: name
-  }
-}
-
-function body(...seq) {
-  return {
-    tag: "UnguardedBody",
-    contents: sequence(...seq)
-  }
-}
-
-
-function entryPoint(name, ...seq) {
-  return {
-    tag: "EntryPoint",
-    contents: [
-      name,
-      sequence(...seq)
-    ]
-  }
-}
-
-function sequence(...seq) {
-  if (seq.length == 0) return none()
-  if (seq.length == 1) return seq[0]
-  return {
-    tag: "Sequence",
-    contents: seq
-  }
-}
-
-function reference(name) {
-  return {
-    tag: "Reference",
-    contents: name
-  }
-}
-
-function application(name, ...params) {
-  return {
-    tag: "Application",
-    contents: [
-      reference(name),
-      params
-    ]
-  }
-}
-
-function repeat(count, ...seq) {
-  return {
-    tag: "Repeat",
-    contents: [
-      count,
-      sequence(...seq)
-    ]
-  }
-}
-
-function muIf(condition, ...seq) {
-  return {
-    tag: "If",
-    contents: [
-      condition,
-      sequence(...seq),
-      none()
-    ]
-  }
-}
-
-function ifElse(condition, seqTrue, seqFalse) {
-  return {
-    tag: "If",
-    contents: [
-      condition,
-      seqTrue,
-      seqFalse
-    ]
-  }
-}
-
-function muWhile(condition, ...seq) {
-  return {
-    tag: "While",
-    contents: [
-      condition,
-      sequence(...seq)
-    ]
-  }
-}
-
-function number(n) {
-  return {
-    tag: "MuNumber",
-    contents: n
-  }
-}
-
-function string(s) {
-  return {
-    tag: "MuString",
-    contents: s
-  }
-}
-
-function none() {
-  return {
-    tag: "None",
-    contents: []
-  }
-}
